+++ stevedore/named.py
@@ -1,7 +1,7 @@
-from .enabled import EnabledExtensionManager
+from .extension import ExtensionManager
 
 
-class NamedExtensionManager(EnabledExtensionManager):
+class NamedExtensionManager(ExtensionManager):
     """Loads only the named extensions.
 
     This is useful for explictly enabling extensions in a
@@ -26,12 +26,20 @@ class NamedExtensionManager(EnabledExtensionManager):
 
     def __init__(self, namespace, names,
                  invoke_on_load=False, invoke_args=(), invoke_kwds={}):
-        def check(ep):
-            return ep.name in names
+        self._names = names
         super(NamedExtensionManager, self).__init__(
             namespace,
-            check,
             invoke_on_load=invoke_on_load,
             invoke_args=invoke_args,
             invoke_kwds=invoke_kwds,
+            )
+
+    def _load_one_plugin(self, ep, invoke_on_load, invoke_args, invoke_kwds):
+        # Check the name before going any further to prevent
+        # undesirable code from being loaded at all if we are not
+        # going to use it.
+        if ep.name not in self._names:
+            return None
+        return super(NamedExtensionManager, self)._load_one_plugin(
+            ep, invoke_on_load, invoke_args, invoke_kwds,
             )